+++ built/jtree.js
@@ -7,6 +7,7 @@ const GrammarBackedNonTerminalNode_1 = require("./grammar/GrammarBackedNonTermin
 const GrammarBackedTerminalNode_1 = require("./grammar/GrammarBackedTerminalNode");
 const GrammarBackedAnyNode_1 = require("./grammar/GrammarBackedAnyNode");
 const GrammarProgram_1 = require("./grammar/GrammarProgram");
+const UnknownGrammarProgram_1 = require("./grammar/UnknownGrammarProgram");
 const TreeNotationCodeMirrorMode_1 = require("./grammar/TreeNotationCodeMirrorMode");
 class jtree {
 }
@@ -17,6 +18,7 @@ jtree.NonTerminalNode = GrammarBackedNonTerminalNode_1.default;
 jtree.TerminalNode = GrammarBackedTerminalNode_1.default;
 jtree.AnyNode = GrammarBackedAnyNode_1.default;
 jtree.GrammarProgram = GrammarProgram_1.default;
+jtree.UnknownGrammarProgram = UnknownGrammarProgram_1.default;
 jtree.TreeNotationCodeMirrorMode = TreeNotationCodeMirrorMode_1.default;
 jtree.getVersion = () => "21.0.0";
 exports.default = jtree;