+++ src/grammar/GrammarProgram.ts
@@ -7,6 +7,7 @@ import { GrammarConstants } from "./GrammarConstants"
 import AbstractGrammarDefinitionNode from "./AbstractGrammarDefinitionNode"
 import GrammarKeywordDefinitionNode from "./GrammarKeywordDefinitionNode"
 import GrammarCellTypeDefinitionNode from "./GrammarCellTypeDefinitionNode"
+import UnknownGrammarProgram from "./UnknownGrammarProgram"
 
 import types from "../types"
 
@@ -274,78 +275,6 @@ ${GrammarConstants.cellType} any`
     })
 
     return new GrammarProgram(tree.getExpanded(1, 2), grammarPath)
-  }
-
-  static _getBestType(values: any) {
-    const all = (fn: Function) => {
-      for (let i = 0; i < values.length; i++) {
-        if (!fn(values[i])) return false
-      }
-      return true
-    }
-    if (all((str: string) => str === "0" || str === "1")) return "bit"
-
-    if (
-      all((str: string) => {
-        const num = parseInt(str)
-        if (isNaN(num)) return false
-        return num.toString() === str
-      })
-    ) {
-      return "int"
-    }
-
-    if (all((str: string) => !str.match(/[^\d\.\-]/))) return "float"
-
-    const bools = new Set(["1", "0", "true", "false", "t", "f", "yes", "no"])
-    if (all((str: string) => bools.has(str.toLowerCase()))) return "bool"
-
-    return "any"
-  }
-
-  static predictGrammarFile(str: string | TreeNode, keywords: string[] = undefined): string {
-    const tree = str instanceof TreeNode ? str : new TreeNode(str)
-    const xi = " " // todo: make param?
-    keywords = keywords || tree.getColumnNames()
-    return keywords //this.getInvalidKeywords()
-      .map((keyword: string) => {
-        const lines = tree.getColumn(keyword).filter(i => i)
-        const cells = lines.map(line => line.split(xi))
-        const sizes = new Set(cells.map(c => c.length))
-        const max = Math.max(...Array.from(sizes))
-        const min = Math.min(...Array.from(sizes))
-        let catchAllCellType: string
-        let cellTypes = []
-        for (let index = 0; index < max; index++) {
-          const set = new Set(cells.map(c => c[index]))
-          const values = Array.from(set).filter(c => c)
-          const type = GrammarProgram._getBestType(values)
-          cellTypes.push(type)
-        }
-        if (max > min) {
-          //columns = columns.slice(0, min)
-          catchAllCellType = cellTypes.pop()
-          while (cellTypes[cellTypes.length - 1] === catchAllCellType) {
-            cellTypes.pop()
-          }
-        }
-
-        const catchAllCellTypeString = catchAllCellType
-          ? `\n ${GrammarConstants.catchAllCellType} ${catchAllCellType}`
-          : ""
-
-        const childrenAnyString = tree.isLeafColumn(keyword) ? "" : `\n ${GrammarConstants.any}`
-
-        if (!cellTypes.length)
-          return `${GrammarConstants.keyword} ${keyword}${catchAllCellTypeString}${childrenAnyString}`
-
-        if (cellTypes.length > 1)
-          return `${GrammarConstants.keyword} ${keyword}
- ${GrammarConstants.cells} ${cellTypes.join(xi)}${catchAllCellTypeString}${childrenAnyString}`
-
-        return `${GrammarConstants.keyword} ${keyword} ${cellTypes[0]}${catchAllCellTypeString}${childrenAnyString}`
-      })
-      .join("\n")
   }
 }